Dam you machine

+++ Assets/_src/Scripts/SplineMovements/SplineMovements.Data/SplineMoveComponent.cs
@@ -7,7 +7,7 @@ namespace _src.Scripts.SplineMovements.SplineMovements.Data
     {
         public byte CurveIndex;
         public half Speed;
-        public half Distance;
+        public float Distance;
         public half Lerp;
     }
 }
